refactor(modal): merge files into a single file
