--- packages/bezier-react/src/components/Modals/Modal/ModalBody.tsx
@@ -1,22 +1,0 @@
-import React, { forwardRef } from 'react'
-
-import { type ModalBodyProps } from './Modal.types'
-
-import * as Styled from './Modal.styled'
-
-/**
- * `ModalBody` is a simple wrapper of the main modal content.
- */
-export const ModalBody = forwardRef(function ModalBody({
-  children,
-  ...rest
-}: ModalBodyProps, forwardedRef: React.Ref<HTMLDivElement>) {
-  return (
-    <Styled.Body
-      ref={forwardedRef}
-      {...rest}
-    >
-      { children }
-    </Styled.Body>
-  )
-})
