--- packages/bezier-react/src/components/Modals/Modal/ModalHelpers.tsx
@@ -1,34 +1,0 @@
-import React from 'react'
-
-import * as DialogPrimitive from '@radix-ui/react-dialog'
-
-import {
-  type ModalCloseProps,
-  type ModalTriggerProps,
-} from './Modal.types'
-
-/**
- * `ModalTrigger` is a button that opens the modal. **It doesn't render any DOM node.**
- * It passes the handler that opens the modal and accessibility properties to the children.
- *
- * It **must** be placed outside of the `ModalContent`.
- */
-export function ModalTrigger({ children }: ModalTriggerProps) {
-  return (
-    <DialogPrimitive.Trigger asChild>
-      { children }
-    </DialogPrimitive.Trigger>
-  )
-}
-
-/**
- * `ModalClose` is a button that closes the modal. **It doesn't render any DOM node.**
- * It passes the handler that closes the modal to the children.
- */
-export function ModalClose({ children }: ModalCloseProps) {
-  return (
-    <DialogPrimitive.Close asChild>
-      { children }
-    </DialogPrimitive.Close>
-  )
-}
